requirements.txt
Django==1.7.5
djangorestframework==3.0.5
djangorestframework-jwt==1.2.0
